src/requirements.txt
strands-agents==1.9.1
boto3>=1.34.0
botocore>=1.34.0
